requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python-headless>=4.5.0
pillow>=9.0.0
tqdm>=4.62.0

# Deep learning
torch>=2.0.0
transformers>=4.30.0
timm>=0.9.2

# Machine learning
scikit-learn>=1.0.0
hmmlearn>=0.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
scipy>=1.10.0
torchvision>=0.16.0
pandas>=2.0.0
openpyxl>=3.1.0

# Optional but recommended
ipython>=8.0.0
jupyter>=1.0.0